231227_전체 페이지 권한 체크

+++ team2/app/Http/Controllers/BoardController.php
@@ -25,12 +25,12 @@ class BoardController extends Controller
      */
     public function index()
     {  
-        $u_id = session('id');
-
         if(!Auth::check()){
-        return redirect()->route('login.get');
+            return redirect()->route('login.get');
         }
-
+        
+        $u_id = session('id');
+        
         $weekAgo = Carbon::now()->subWeek();
         
         $hotboard = Board::orderBy('board_hits', 'desc')
@@ -85,6 +85,10 @@ class BoardController extends Controller
     }
 
     public function categoryboard(){
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $category_board=Board::where('category_id', '1',)->orderBy('board_id', 'desc')->paginate(5);
         $category_id = Category::orderby('category_id', 'asc')->get();
         $category_name = Category::where('category_id', '1')->get();
@@ -102,6 +106,10 @@ class BoardController extends Controller
      */
     public function create()
     {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $result= Hashtag::all();
         return view('insert')->with('data', $result);
     }    
@@ -113,7 +121,10 @@ class BoardController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {      
+    {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
 
         $u_id = auth()->id();        
         $boardData = $request->only('board_title', 'board_content', 'category_id');
@@ -184,6 +195,10 @@ class BoardController extends Controller
      */
     public function show($board_id)
     {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $result = Board::with(['user', 'images'])->find($board_id);
 
         //  dd($result->images);
@@ -202,11 +217,13 @@ class BoardController extends Controller
      */
     public function edit($board_id)
     {
-        
+        if(!Auth::check()){
+        return redirect()->route('login.get');
+        }
 
-    $result = Board::find($board_id);
-    $allHashtags = Hashtag::all();
-    return view('update', compact('result', 'allHashtags'));
+        $result = Board::find($board_id);
+        $allHashtags = Hashtag::all();
+        return view('update', compact('result', 'allHashtags'));
 
     }
 
@@ -219,6 +236,10 @@ class BoardController extends Controller
      */
     public function update(Request $request, $board_id)
     {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $result = Board::find($board_id);
         $result->update([
             'board_title' => $request->input('u_title'),
@@ -285,11 +306,20 @@ $board_detail_get = Board::with(['hashtags'])
      */
     public function destroy($board_id)
     {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         Board::destroy($board_id);
         return redirect()-> route('categoryboard');
     }
 
     public function boardcategoryget($categoryId) {
+
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+            
         // Log::debug($categoryId);
         $category_board = Board::where('category_id', $categoryId)->orderby('board_id', 'desc')->paginate(5);
 
@@ -349,18 +379,30 @@ $board_detail_get = Board::with(['hashtags'])
     }
 
     public function lastboardget() {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $lastboard = Board::orderBy('board_id', 'desc')->paginate(5);
 
         return view('lastboard')->with('data', $lastboard);
     }
 
     public function hotboardget() {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $hotboard = Board::orderBy('board_hits', 'desc')->paginate(5);
 
         return view('hotboard')->with('data', $hotboard);
     }
 
     public function favoriteboardget() {
+        if(!Auth::check()){
+            return redirect()->route('login.get');
+            }
+
         $u_id = session('id');
 
         $favoriteboard = User::join('favorite_tags', 'users.id', '=', 'favorite_tags.u_id')
